+++ klaus/contrib/app_args.py
@@ -26,6 +26,5 @@ def get_args_from_env():
         unauthenticated_push=strtobool(
             os.environ.get("KLAUS_UNAUTHENTICATED_PUSH", "0")
         ),
-        ctags_policy=os.environ.get("KLAUS_CTAGS_POLICY", "none"),
     )
     return args, kwargs